[PropertyInfo] Extract the logic converting a php doc to a Type in PhpDocTypeHelperTrait
